refactor: move PrizeConstants to its own package `domain.prizes`
