+++ musical-surveyor-app/src/main/java/io/davorpatech/apps/musicalsurveyor/persistence/model/Prize.java
@@ -1,6 +1,6 @@
 package io.davorpatech.apps.musicalsurveyor.persistence.model;
 
-import io.davorpatech.apps.musicalsurveyor.domain.PrizeConstants;
+import io.davorpatech.apps.musicalsurveyor.domain.prizes.PrizeConstants;
 import io.davorpatech.fwk.auditing.jpa.Audit;
 import io.davorpatech.fwk.auditing.jpa.AuditAccessor;
 import io.davorpatech.fwk.model.BaseEntity;
